requirements.txt
django-environ,
djangorestframework,
django-filter,